% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignClusters.R
\name{align_clusters_list}
\alias{align_clusters_list}
\title{Align multiple clusterings in list format}
\usage{
align_clusters_list(
  clust_list,
  clust_ref,
  clust_names = NULL,
  handle_unmatched = "label",
  verbose = FALSE
)
}
\description{
Align multiple clusterings in list format
}
\keyword{internal}
