% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignClusters.R
\name{align_single_clustering}
\alias{align_single_clustering}
\title{Align a single clustering to reference}
\usage{
align_single_clustering(clust, clust_ref, handle_unmatched = "label")
}
\description{
Internal function for aligning a single clustering vector
}
\keyword{internal}
