% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{coerce_ordinal_numeric}
\alias{coerce_ordinal_numeric}
\title{Coerce Ordinal Labels to Numeric (e.g. "1", "1 or 2")}
\usage{
coerce_ordinal_numeric(x)
}
\arguments{
\item{x}{Character or factor vector}
}
\value{
List with fields: numeric (numeric vector) and success (logical)
}
\description{
Attempts to convert factor/character ordinal labels to numeric values.
Recognises patterns like "a or b" and uses their average.
}
\keyword{internal}
