% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get Color Palette}
\usage{
get_color_palette(palette_name = "matlab")
}
\arguments{
\item{palette_name}{Character. Name of the palette.}
}
\value{
Character vector of colors.
}
\description{
Internal function to get color palettes for continuous variables.
}
\keyword{internal}
