% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{get_discrete_colors}
\alias{get_discrete_colors}
\title{Get Discrete Color Palette}
\usage{
get_discrete_colors(level_names, palette_name = "matlab", reverse = FALSE)
}
\arguments{
\item{level_names}{Character vector of level names}

\item{palette_name}{Name of the palette}

\item{reverse}{Logical. If TRUE, reverse the color order}
}
\value{
Named vector of colors
}
\description{
Get Discrete Color Palette
}
\keyword{internal}
