% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{is_discrete_variable}
\alias{is_discrete_variable}
\title{Detect if Variable Should Use Discrete Coloring}
\usage{
is_discrete_variable(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical indicating whether to use discrete colors
}
\description{
Detect if Variable Should Use Discrete Coloring
}
\keyword{internal}
