% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{prepare_discrete_coloring}
\alias{prepare_discrete_coloring}
\title{Prepare Discrete Coloring Variable}
\usage{
prepare_discrete_coloring(x, sort_levels = TRUE, reverse = FALSE)
}
\arguments{
\item{x}{Vector to prepare}

\item{sort_levels}{Whether to sort factor levels}

\item{reverse}{Whether to reverse color gradient}
}
\value{
List with prepared factor and color mapping
}
\description{
Prepare Discrete Coloring Variable
}
\keyword{internal}
